Added inputLayer class
This class initialises an input layer and holds the input neuron values.
Other functions can then access and set the input neuron values through
this class. This class was just proof of concept, to make sure I
understood how to move from procedural to object oriented programming,
from here I can now add class for the other layers, weight, errors, etc,
with the aim of making the program as modular and parameterised as
possible.  Once I have achieved this it will just be a matter of
programming interfaces and tweaking the network settings.