Two Changes
I'm going to commit two changes here, which is probably bad form as far
as version control goes, but it all works.

The first changes I've made is to add the files, 'Network.cpp' and
'Network.h', these files are going to contain the information about the
classes from now on. I'm gradually going to move all the functions
currently in 'main.cpp' over to separate files to stop it getting
cluttered.

The second change I've made is to add a hiddenLayer class, and write it
in to the main code. However it occurred to me whilst I was writing it
that I  only actually need one layer class, and just create many
instances of the same class. I'm going to keep going with the way I'm
doing it now, but this should be implemented in future.