Added outputLayer class
The outputLayer class was output, so now all the neuron values are
stored in class. I had some more thoughts on how to implement all the
layers as multiple instances of one class, I need to add an additional
parameter to the constructor, just a boolean, so that the addition of a
bias neuron is true or false.

The next thing I need to do is create the classes which hold all the
errors. This is probably going to be slightly more complicated as the
change to object oriented is going to be slightly more radical. I'm
going to move to creating all errors as instances of one error class.