Added weights class
Weights are now contained in a class in a similar way to layers, however
unlike the layer class, there is only one type of weights class, which
is how the layer class should of been implemented, but now it must be
changed in a later version.

Now I need to make a training class, which I think I'm going to set up
as a different *.cpp and *.h file, so that when I have different types
of trainers, networks, and neurons, everything will be collect together.