Moved layer calculations
The layer calculation has been moved from a function from the main
program, to being contain within the layer class. The compiler had to be
told that the weights class existed in the Network.h file in order to
avoid a circular dependency issue. Next I need to move the update
weights class to within the weights function.