Pretty much there
I believe that all of the network code is now object-oriented. The code
that isn't is all for data input, which I suppose could be changed to
object-oriented fairly quickly. What I need to do know is find out how
to pass arrays as parameter to instances, instead of the ugly code, and
then start tiding everything up, like getting rid of all the checking
functions I wrote when developing the original program. I think I'll
just move them into a seperate C++ file, and just not include it in the
project,  but still in the git repository.